Exclude full file name link from nowrap to improve display
Long filenames, such as those generated by the filename templates, can cause display issues such as pushing the submission details table (teacher view) beyond the content area of the page.  It's not too bad in the list of submissions if the icon and embedded player are in the nowrap but the full file name isn't.